requirements.txt
numpy
opencv-python
python-telegram-bot
SpeechRecognition
